Rename scanner modules
