+++ scanner/Dockerfile
@@ -5,4 +5,5 @@ COPY ./requirements.txt .
 RUN pip3 install -r ./requirements.txt
 
 COPY . .
-ENTRYPOINT ["python3", "-m", "scanner"]
+ENTRYPOINT ["python3", "-m"]
+CMD ["scanner"]
