+++ scanner/requirements.txt
@@ -3,3 +3,4 @@ aiohttp
 jsons
 watchfiles
 aio-pika
+dataclasses-json
